backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic[email]
email-validator
